yo
git-svn-id: https://mason.googlecode.com/svn/trunk@1016 e22d1478-1c87-2d72-604f-1c9cf2634165

+++ mason/sim/portrayal/grid/FastObjectGridPortrayal2D.java
@@ -134,7 +134,7 @@ public class FastObjectGridPortrayal2D extends ObjectGridPortrayal2D
     	If FRACTION or MODULUS is set to a value <= 0, or COLOR is set to null, then the grid is not displayed.
     	If you can't think of a good FRACTION, 1/8.0 doesn't look too bad.
     */
-    public void setGridLines(float fraction, int modulus, Color color)
+    public void setGridLines(double fraction, int modulus, Color color)
     	{
     	valueGridPortrayal.setGridLines(fraction, modulus, color);
     	}
